refactor: TicketEntryTime 검증 로직 Ticket으로 이동
